Move icon observer to marker-layer
